requirements.txt
bottle-0.12.25